feat: страница edit/id переведена на ssr
