+++ packages/components/index.js
@@ -32,3 +32,4 @@ export { default as IndexBlock } from './src/blocks/index-block/index-block.jsx'
 export { default as TestIdBlock } from './src/blocks/test-id-block/test-id-block.jsx';
 export { default as EditBlock } from './src/blocks/edit-block/edit-block.jsx';
 export { default as EditCreateBlock } from './src/blocks/edit-create-block/edit-create-block.jsx';
+export { default as EditIdBlock } from './src/blocks/edit-id-block/edit-id-block.jsx';
